seperated the which file into path file
