+++ __exit.c
@@ -2,7 +2,7 @@
 
 
 /**
- * _atoi - convert string to int
+ * c_atoi - convert string to int
  *
  * @s: Value of input of string
  *
